Refactor Export-Registry
'Export-Registry.ps1 was refactored to move the formatting into the dot include file 'Format-RegExport.ps1'. Additional comments were added to document how the scripts work. A registry test file was added to verify the import and export of all the supported types creates character-by-character and byte-by-byte identical match to a RegEdit exported registry file of the same key for the supported value types.